refactor into transfer panel
