+++ src/components/SelectAssetModal/index.ts
@@ -1,0 +1,3 @@
+import SelectAssetModal from './SelectAssetModal';
+
+export default SelectAssetModal;
