--- src/components/SelectTokenModal/index.ts
@@ -1,3 +1,0 @@
-import SelectTokenModal from './SelectTokenModal';
-
-export default SelectTokenModal;
